move histowiki api into its own file
